pyproject.toml
[project]
name = "nas-orchestrator"
version = "0.1.0"
description = "Automated NAS stack orchestrator"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Ethan", email = "ethan@example.com" }
]
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.10",
  "pydantic-settings>=2.5",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "httpx>=0.27",
  "sse-starlette>=2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.21",
  "ruff>=0.1",
  "mypy>=1.9"
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["orchestrator"]
